[ncurses] Update to 6.3. JB#55136
The most interesting part of this update was the libtinfo deletion
which had got broken on commit 622c5430ddf6 due to deleting non-existing
libtinfo.6* instead of libtinfo.so.6*. As result apps using ncursesw
were commonly segfaulting on startup, suppose from having a wide version
of the library and narrow version of the libtinfo. Why the previous
version was working, maybe it was just luck?

Overall the setup of one common libtinfo for both the variants seems
crazy, but seems Fedora is also doing that. Considered droppping the
common name for both variants, but could be problematic for updates
due to existing low-level packages depending on libtinfo.

Getting rid of the narrow versions altogether could be an option.